Make way for ARM, move existing code into smapi.
Follow the conventions used by packer, and subdivide the directory into
the various packers.  The existing code uses the Service Management API,
and will now live under smapi.  The ARM based packer will eventually live
under arm.

This is just the rename operation because there are quite a few edits that
need to be made to fix the build.

Move utils to a common area.
Move constants to the common directory.
Move the lin package to the common area.
