+++ packer/builder/azure/common/constants/goos.go
