+++ packer/builder/azure/common/constants/stateBag.go
