+++ packer/builder/azure/common/constants/targetplatforms.go
