+++ packer/builder/azure/common/lin/ssh.go
@@ -6,7 +6,7 @@ package lin
 
 import (
 	"fmt"
-	"github.com/Azure/packer-azure/packer/builder/azure/constants"
+	"github.com/Azure/packer-azure/packer/builder/azure/common/constants"
 	"github.com/mitchellh/multistep"
 	"golang.org/x/crypto/ssh"
 )
