+++ packer/builder/azure/common/lin/step_create_cert.go
@@ -16,7 +16,7 @@ import (
 	"math/big"
 	"time"
 
-	"github.com/Azure/packer-azure/packer/builder/azure/constants"
+	"github.com/Azure/packer-azure/packer/builder/azure/common/constants"
 
 	"github.com/mitchellh/multistep"
 	"github.com/mitchellh/packer/packer"
