+++ packer/builder/azure/common/lin/step_generalize_os.go
