+++ packer/builder/azure/common/utils/gluestrings.go
