+++ packer/builder/azure/common/utils/gluestrings_test.go
