+++ packer/builder/azure/common/utils/randomstring.go
