+++ packer/builder/azure/common/utils/randomstring_test.go
