+++ packer/builder/azure/smapi/FindImage.go
