+++ packer/builder/azure/smapi/FindImage_test.go
