+++ packer/builder/azure/smapi/artifact.go
