+++ packer/builder/azure/smapi/config_test.go
