+++ packer/builder/azure/smapi/publishsettings.go
