+++ packer/builder/azure/smapi/publishsettings_test.go
