+++ packer/builder/azure/smapi/requestlogger.go
