+++ packer/builder/azure/smapi/retry/operations.go
