+++ packer/builder/azure/smapi/step_create_vm.go
@@ -7,8 +7,8 @@ package azure
 import (
 	"fmt"
 
-	"github.com/Azure/packer-azure/packer/builder/azure/constants"
-	"github.com/Azure/packer-azure/packer/builder/azure/retry"
+	"github.com/Azure/packer-azure/packer/builder/azure/common/constants"
+	"github.com/Azure/packer-azure/packer/builder/azure/smapi/retry"
 
 	"github.com/mitchellh/multistep"
 	"github.com/mitchellh/packer/packer"
