+++ packer/builder/azure/smapi/step_poll_status.go
@@ -9,7 +9,7 @@ import (
 	"log"
 	"time"
 
-	"github.com/Azure/packer-azure/packer/builder/azure/constants"
+	"github.com/Azure/packer-azure/packer/builder/azure/common/constants"
 
 	"github.com/mitchellh/multistep"
 	"github.com/mitchellh/packer/packer"
