+++ packer/builder/azure/smapi/step_validate.go
@@ -9,8 +9,8 @@ import (
 	vmdisk "github.com/Azure/azure-sdk-for-go/management/virtualmachinedisk"
 	vmimage "github.com/Azure/azure-sdk-for-go/management/virtualmachineimage"
 	"github.com/Azure/azure-sdk-for-go/management/vmutils"
-	"github.com/Azure/packer-azure/packer/builder/azure/constants"
-	"github.com/Azure/packer-azure/packer/builder/azure/utils"
+	"github.com/Azure/packer-azure/packer/builder/azure/common/constants"
+	"github.com/Azure/packer-azure/packer/builder/azure/common/utils"
 	"github.com/mitchellh/multistep"
 	"github.com/mitchellh/packer/packer"
 	"log"
